gotstyle-server/requirements.txt
alembic==1.11.1
APScheduler==3.10.4
bcrypt==4.0.1
blinker==1.6.2
boto3==1.28.17
botocore==1.31.17
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
coverage==7.2.7
Flask==2.3.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
gunicorn==21.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
Mako==1.2.4
MarkupSafe==2.1.3
mysqlclient==2.2.0
packaging==23.1
Pillow==10.0.0
psycopg2==2.9.6
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
s3transfer==0.6.1
six==1.16.0
SQLAlchemy==2.0.19
typing_extensions==4.7.1
tzlocal==5.0.1
urllib3==1.26.16
Werkzeug==2.3.6
